Live2D_digital_human/awesome-digital-human-live2d/digitalHuman/crawler/requirements.txt
# 核心依赖
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# 数据库
sqlalchemy>=2.0.0
pymysql>=1.1.0

# 数据处理
pandas>=2.0.0

# 定时任务
apscheduler>=3.10.0

# 工具库
python-dotenv>=1.0.0